Add VramConfiguration helpers to render/
The idea here is to allow specifying named configuration sets, which
configure all the banks at once and house references to relevant
addresses and offsets. Currently this is limited to the four "main"
banks used during 3D rendering, and the LCDC offsets that must be
tracked depending on their configuration.

The basic idea is to collect the VramConfiguration settings into
this one location, and retrieve its components which are currently
scattered about through the codebase. In addition to improving
readability, this will make these allocations easier to adjust
moving forward.