Test basic expectations for vram_configuration_for_pass
This set mostly confirms that the even/odd bank selections actually
differ between passes, and that the final render ends up somewhere
that won't be clobbered by the next partial run. We also confirm
that LCDC addresses map appropriately, and that any previous cature
data is mapped in as texture memory.

Right now the tests assume that Slot 0 will be used for the LCDC
target. I don't like hard-coding this though, and I think we should
generalize the texture slot selection and track the offset within
the VramConfiguration structure.